requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
scikit-image>=0.20.0
trimesh>=3.20.0
Pillow>=9.5.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0
pyyaml>=6.0
opencv-python>=4.7.0
einops>=0.6.0
transformers>=4.30.0
timm>=0.9.0
pyrender>=0.1.45
imageio>=2.30.0
pytest>=7.4.0
pytest-cov>=4.1.0

